+++ src/main/java/fll/web/schedule/CheckViolations.java
@@ -20,7 +20,9 @@ import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
 import javax.sql.DataSource;
 
+import fll.Team;
 import fll.db.Queries;
+import fll.db.TournamentParameters;
 import fll.scheduler.ConstraintViolation;
 import fll.scheduler.SchedParams;
 import fll.scheduler.ScheduleChecker;
@@ -71,6 +73,15 @@ public class CheckViolations extends BaseFLLServlet {
       final SchedParams schedParams = uploadScheduleData.getSchedParams();
       final ScheduleChecker checker = new ScheduleChecker(schedParams, schedule);
       violations.addAll(checker.verifySchedule());
+
+      final int numSeedingRounds = TournamentParameters.getNumSeedingRounds(connection, tournamentID);
+      if (uploadScheduleData.getNumPerformanceRuns() < numSeedingRounds) {
+        violations.add(new ConstraintViolation(ConstraintViolation.Type.HARD, Team.NULL_TEAM_NUMBER, null, null, null,
+                                               String.format("number of performance (%d) runs must be greater than or equal to the number of seeding rounds (%d)",
+                                                             uploadScheduleData.getNumPerformanceRuns(),
+                                                             numSeedingRounds)));
+      }
+      
       if (violations.isEmpty()) {
         WebUtils.sendRedirect(application, response, "/schedule/GatherTeamInformationChanges");
         return;